CDI migration, cart persistence, payment flow, pro tarif wizard — large batch
- CDI migration: rowmappers extracted to dedicated classes, CrudService removed,
  GenericDAO streamlined, entite POJOs cleaned up
- Cart persistence: DB-only cart model (CartStatus enum, CartRowMapper,
  UpdateCartStatus, DeleteCart, FindCart, cart_migration SQL)
- Payment flow: eTypePayment enum, PaymentOrchestrator, LessonRegistrar,
  PaymentLessonController (professional null guard), SubscriptionRegistrar
- Pro tarif wizard: ProTarif entity, CalcLessonPrice, tarif_professional_wizard.xhtml,
  include_tarif_pro_period/pricing, FindLessonBooked, UpdateProfessional
- New finders: FindCotisationOverlapping, FindSubscriptionOverlapping, FindGreenfeePaid
- XHTML: ground_condition_display, greenfee_equipment, cart includes, menu, header,
  selectPro, selectClubCourse, schedule updates
- i18n: messages_fr properties updates
- Tests: DeleteClubIT, HttpControllerTest, arquillian.xml updated
- Removed: CardBrand enum (replaced by CreditcardBrand), CrudService, cart_pro.xhtml

+++ src/main/java/mail/LessonMail.java
@@ -38,7 +38,7 @@ public class LessonMail implements Serializable {
 
             String cs = currSymbol(creditcard);
             double lessonTotal = lessons.stream()
-                .mapToDouble(l -> l.getLessonAmount() != null ? l.getLessonAmount() : 0.0)
+                .mapToDouble(l -> l.getLessonAmount() != null ? l.getLessonAmount().doubleValue() : 0.0)
                 .sum();
             StringBuilder lignes = new StringBuilder();
             for (Lesson lesson : lessons) {
@@ -47,7 +47,7 @@ public class LessonMail implements Serializable {
                 lignes.append("<tr><td>🎓 ").append(lesson.getEventTitle()).append("</td>")
                       .append("<td style='color:#555'>").append(start).append(" → ").append(end).append("</td>")
                       .append("<td align='right'>")
-                      .append(String.format("%.2f %s", lesson.getLessonAmount() != null ? lesson.getLessonAmount() : 0.0, cs))
+                      .append(String.format("%.2f %s", lesson.getLessonAmount() != null ? lesson.getLessonAmount().doubleValue() : 0.0, cs))
                       .append("</td></tr>");
             }
 
@@ -96,7 +96,7 @@ public class LessonMail implements Serializable {
 
             String clubName = !lessons.isEmpty() ? lessons.get(0).getEventClubName() : "";
             double lessonTotal = lessons.stream()
-                .mapToDouble(l -> l.getLessonAmount() != null ? l.getLessonAmount() : 0.0)
+                .mapToDouble(l -> l.getLessonAmount() != null ? l.getLessonAmount().doubleValue() : 0.0)
                 .sum();
 
             StringBuilder lignes = new StringBuilder();
